+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Changed view engine to use `Map<String, dynamic>`.
 * Remove dependency on `package:json_god` by default.
 * Remove dependency on `package:dart2_constant`.
+* Remove `contentType` argument in `ResponseContext.serialize`.
 
 # 1.1.5+1
 * Patched annoying error that prevented MapServices from working,